@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Map;
 
 import cn.jpush.im.android.api.JMessageClient;
+import cn.jpush.im.android.api.model.Conversation;
 import cn.jpush.im.android.api.model.GroupInfo;
 import cn.jpush.im.android.api.model.Message;
 import cn.jpush.im.android.api.model.UserInfo;
@@ -76,14 +77,19 @@ public class JGApplication extends com.activeandroid.app.Application {
     public static final int REQUEST_CODE_SEND_LOCATION = 24;
     public static final int REQUEST_CODE_FRIEND_INFO = 16;
     public static final int RESULT_CODE_CHAT_DETAIL = 15;
+    public static final int REQUEST_CODE_FRIEND_LIST = 17;
     public static final int ON_GROUP_EVENT = 3004;
     public static final String DELETE_MODE = "deleteMode";
     public static final int RESULT_CODE_ME_INFO = 20;
 
     public static final String DRAFT = "draft";
+    public static final String CONV_TYPE = "conversationType"; //value使用 ConversationType
+    public static final String ROOM_ID = "roomId";
     public static final String GROUP_ID = "groupId";
     public static final String POSITION = "position";
     public static final String MsgIDs = "msgIDs";
+    public static final String MSG_JSON = "msg_json";
+    public static final String MSG_LIST_JSON = "msg_list_json";
     public static final String NAME = "name";
     public static final String ATALL = "atall";
     public static final String SEARCH_AT_MEMBER_NAME = "search_at_member_name";
@@ -96,11 +102,13 @@ public class JGApplication extends com.activeandroid.app.Application {
     private static final String JCHAT_CONFIGS = "JChat_configs";
     public static String FILE_DIR = "sdcard/JChatDemo/recvFiles/";
     public static String VIDEO_DIR = "sdcarVIDEOd/JChatDemo/sendFiles/";
+    public static String THUMP_PICTURE_DIR;
     public static final String TARGET_ID = "targetId";
     public static final String ATUSER = "atuser";
     public static final String TARGET_APP_KEY = "targetAppKey";
     public static int maxImgCount;               //允许选择图片最大数
     public static final String GROUP_NAME = "groupName";
+    public static String groupAvatarPath;
 
     public static Context context;
     public static LocationService locationService;
@@ -113,13 +121,16 @@ public class JGApplication extends com.activeandroid.app.Application {
     public static List<UserInfo> alreadyRead = new ArrayList<>();
     public static List<UserInfo> unRead = new ArrayList<>();
     public static List<String> forAddFriend = new ArrayList<>();
+    public static List<String> forAddIntoGroup = new ArrayList<>();
+    public static Conversation delConversation;
+    public static ArrayList<String> selectedUser;
 
     @Override
     public void onCreate() {
         super.onCreate();
         context = getApplicationContext();
+        THUMP_PICTURE_DIR = context.getFilesDir().getAbsolutePath() + "/JChatDemo";
         StorageUtil.init(context, null);
-
         Fresco.initialize(getApplicationContext());
         SDKInitializer.initialize(getApplicationContext());
         locationService = new LocationService(getApplicationContext());
