@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.support.v4.app.FragmentActivity;
 import android.text.TextUtils;
 import android.view.View;
+import android.widget.TextView;
 
 import com.activeandroid.ActiveAndroid;
 
@@ -17,10 +18,10 @@ import cn.jpush.im.android.api.JMessageClient;
 import cn.jpush.im.android.api.callback.GetUserInfoListCallback;
 import cn.jpush.im.android.api.model.UserInfo;
 import jiguang.chat.R;
-import jiguang.chat.activity.FriendRecommendActivity;
 import jiguang.chat.activity.GroupActivity;
 import jiguang.chat.activity.SearchContactsActivity;
 import jiguang.chat.activity.SearchForAddFriendActivity;
+import jiguang.chat.activity.VerificationMessageActivity;
 import jiguang.chat.adapter.StickyListAdapter;
 import jiguang.chat.application.JGApplication;
 import jiguang.chat.database.FriendEntry;
@@ -39,12 +40,14 @@ public class ContactsController implements View.OnClickListener, SideBar.OnTouch
     private Activity mContext;
     private List<FriendEntry> mList = new ArrayList<>();
     private StickyListAdapter mAdapter;
+    private TextView mAllContactNumber;
     private List<FriendEntry> forDelete = new ArrayList<>();
 
 
     public ContactsController(ContactsView mContactsView, FragmentActivity context) {
         this.mContactsView = mContactsView;
         this.mContext = context;
+        mAllContactNumber = mContext.findViewById(R.id.all_contact_number);
     }
 
 
@@ -57,9 +60,11 @@ public class ContactsController implements View.OnClickListener, SideBar.OnTouch
                 mContext.startActivity(intent);
                 break;
             case R.id.verify_ll://验证消息
-                intent.setClass(mContext, FriendRecommendActivity.class);
+                //intent.setClass(mContext, FriendRecommendActivity.class);
+                intent.setClass(mContext, VerificationMessageActivity.class);
                 mContext.startActivity(intent);
                 mContactsView.dismissNewFriends();
+                mAllContactNumber.setVisibility(View.GONE);
                 break;
             case R.id.group_ll://群组
                 intent.setClass(mContext, GroupActivity.class);
@@ -84,7 +89,7 @@ public class ContactsController implements View.OnClickListener, SideBar.OnTouch
             public void gotResult(int responseCode, String responseMessage, List<UserInfo> userInfoList) {
                 mContactsView.dismissLoadingHeader();
                 if (responseCode == 0) {
-                    if (userInfoList.size() != 0) {
+                    if (userInfoList != null && userInfoList.size() != 0) {
                         mContactsView.dismissLine();
                         ActiveAndroid.beginTransaction();
                         try {
