@@ -2,13 +2,14 @@ package jiguang.chat.utils;
 
 
 import android.os.Handler;
+import android.os.Looper;
 
 /**
  * Created by ${chenyn} on 2017/3/10.
  */
 
 public class ThreadUtil {
-    static Handler mHandler = new Handler();
+    static Handler mHandler = new Handler(Looper.getMainLooper());
 
     public static void runInThread(Runnable task) {
         new Thread(task).start();
